rename: command_dispatch → command_dispatcher
