+++ CMakeLists.txt
@@ -124,7 +124,7 @@ add_subdirectory(shaders)
 
 add_library(gseurat_core OBJECT
     src/engine/app_base.cpp
-    src/engine/command_dispatch.cpp
+    src/engine/command_dispatcher.cpp
     src/engine/vk_context.cpp
     src/engine/swapchain.cpp
     src/engine/render_pass.cpp
