+++ include/gseurat/engine/command_dispatcher.hpp
